order-service/requirements.txt
Flask==2.3.3
requests==2.32.5
pytest==8.4.2
pytest-mock==3.14.0
pyyaml==6.0.2